Group

+++ src/pages/Project/ProjectList.js
@@ -1,4 +1,4 @@
-import React, {useContext, useEffect, useRef, useState} from "react";
+import React, {useEffect, useRef, useState} from "react";
 import {ProjectService} from "../../services/services";
 import {useLoading} from "../../component/hooks/hooks";
 import FullscreenLoading from "../../component/FullScreenLoading";
@@ -8,73 +8,65 @@ import useStyles from "./styles/listStyle";
 import EditOutlinedIcon from '@material-ui/icons/EditOutlined';
 import DeleteSweepOutlinedIcon from '@material-ui/icons/DeleteSweepOutlined'
 import AddIcon from '@material-ui/icons/Add';
-import DialogModal from "../../component/DialogModal";
 import ProjectCreateModal from "./ProjectCreateModal";
 import moment from "moment";
-import ContextProvider from "../../component/ContextProvider";
+
 import BackButton from "../../component/BackButton";
 import Button from "@material-ui/core/Button";
+import Linker from "../../component/Linker";
+import {useSnackbar} from "notistack";
+import {useConfirm} from "material-ui-confirm";
 
 const ProjectList = () => {
     const [projects, setProjects] = useState([]);
-    const [isShowing, setIsShowing] = useState(false);
+
     const [isCreate, setIsCreate] = useState(false);
     const {loading, onLoading, offLoading} = useLoading();
-    const [id, setId] = useState("");
     const modalRef = useRef(null);
     const classes = useStyles();
     const [mounted, setMounted] = useState(true);
+    const confirm = useConfirm()
+
+    const {enqueueSnackbar} = useSnackbar();
 
     const loadProjects = (e) => {
         setProjects(e);
-    }
-
-    const toggleDelete = () => {
-        setIsShowing(!isShowing);
     }
 
     const toggleCreate = () => {
         setIsCreate(!isCreate);
     }
 
-    const loadId = (e) => {
-        debugger;
-        setId(e)
-        handleOnYes(e);
-    }
     const toggleMount = () => setMounted(!mounted);
 
-    const fetchProject = async () => {
-        await ProjectService.getList()
-            .then((r) => {
-                if (r.status === 200) {
-                    loadProjects(r.data);
-                    console.log(r.data);
-                } else loadProjects([]);
-            }).catch((r) => {
-            console.log(r);
-        });
-    }
-
-    const handleOnYes = async(id) => {
+    const handleOnYes = (id) => {
         onLoading();
-        await ProjectService.deleteProject(id)
-            .then((r) => {
-                if (r.status === 200) {
-                    let index = projects.indexOf(id);
-                    projects.splice(index, 1);
-                    console.log("remove " + id);
-                } else console.log(r);
-            }, null)
+        confirm({description:"Are you sure? This action is permanent."})
+            .then(() =>{
+                ProjectService.deleteProject(id)
+                    .then((r) => {
+                        if (r.status === 200) {
+                            toggleMount();
+                            enqueueSnackbar("Removed success", {variant:"success"})
+                        } else  enqueueSnackbar(r.data.message, {variant:"warning"})
+                    }, null)
+                    .catch((r) => enqueueSnackbar(r, {variant:"error"}))
+            })
         offLoading();
     }
     useEffect(() => {
         onLoading();
-        fetchProject();
+        ProjectService.getList()
+            .then((r) => {
+                if (r.status === 200) {
+                    loadProjects(r.data);
+                } else loadProjects([]);
+            }).catch((r) => {
+            enqueueSnackbar(r, {variant:"error"})
+        });
         offLoading();
+        document.title = "Project List";
         }, [mounted,setMounted]);
-
-    const {switchToEditPro} = useContext(ContextProvider);
 
     const columns = [
         {field: 'name', headerName: 'Project Name', width: 200},
@@ -82,42 +74,25 @@ const ProjectList = () => {
         {
             field: 'startDate', headerName: 'Start Date', width: 200,
             valueFormatter: (params) => {
-                const valueFormatted = moment(params.value).format("Do MMM YYYY");
-                return valueFormatted;
+                return moment(params.value).format("Do MMM YYYY");
             }
         },
         {field: 'dueDate', headerName: 'End Date', width: 200,
             valueFormatter: (params) => {
-                const valueFormatted = moment(params.value).format("Do MMM YYYY");
-                return valueFormatted;
+                return moment(params.value).format("Do MMM YYYY");
             }},
         {
             field: 'action', headerName: 'Actions', width: 200,
             renderCell: (p) => {
                 return (
                     <div>
-                        <DialogModal
-                            isOpened={isShowing}
-                            toggle={toggleDelete}
-                            contents={"Are you sure that you want to delete this?"}
-                            heading={"Delete confirmation"}
-                            onYes={() => {
-                                loadId(p.row.id);
-                            }}
-                        />
-                        <div className={classes.button}>
-                            <EditOutlinedIcon
-                                className={classes.icon}
-                                onClick={() => {
-                                    switchToEditPro(p.row.id);
-                                    console.log(p.row.id);
-                                }}
-                            />
+                        <Linker to={"project/" + p.row.id} content={<EditOutlinedIcon/>}/>
+                        <Button>
                             <DeleteSweepOutlinedIcon
-                                className={classes.icon}
-                                onClick={toggleDelete}
+                                color="secondary"
+                                onClick={() =>handleOnYes(p.row.id)}
                             />
-                        </div>
+                        </Button>
                     </div>
                 )
             }
@@ -125,28 +100,26 @@ const ProjectList = () => {
     ];
 
     return (
-        <div>
+        <>
             {loading ? <FullscreenLoading/> : null}
             <Paper className={classes.root}>
                 <ProjectCreateModal
                     modalRef={modalRef}
                     toggleModal={toggleCreate}
                     isShowing={isCreate}
-                    onReload={fetchProject}
                     toggleMount={toggleMount}
                 />
-                <Grid container justify="center" spacing={3}>
+                <Grid container justify="center" spacing={3} className={classes.content}>
                     <Grid item xs={1}>
                         <Typography variant="h6" align="center">PROJECTS</Typography>
                     </Grid>
                 </Grid>
-                <Grid container justify="center" >
-                    <Grid item xs={3} spacing={2}>
-                        <BackButton children="Back to home"/>
+                <Grid container justifyContent={"space-between"} >
+                    <Grid item xs={3}>
+                        <BackButton children="Back to home" switchTo={"/"}/>
                     </Grid>
-                    <Grid item xs={7}/>
                     <Grid item xs={2}>
-                        <Button onClick={toggleCreate} >
+                        <Button onClick={toggleCreate}>
                             <AddIcon/> Create
                         </Button>
                     </Grid>
@@ -162,7 +135,7 @@ const ProjectList = () => {
                     </Grid>
                 </Grid>
             </Paper>
-        </div>
+        </>
     )
 }
 export default ProjectList
